BotTelegram/requirements.txt
Flask==0.11
gunicorn==19.6
python-telegram-bot
pymongo
